refactor(projection): extract inherit token logic (#68)
